backend/pyproject.toml
[project]
name = "podium"
version = "0.1.0"
description = ""
authors = [
    {name = "slashtechno", email = "77907286+slashtechno@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.5,<0.116.0)",
    "uvicorn[standard] (>=0.32.1,<0.33.0)",
    "dynaconf (>=3.2.6,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sentry-sdk[fastapi] (>=2.22.0,<3.0.0)",
    "httpx>=0.28.1,<0.29.0",
    "python-slugify>=8.0.4,<9.0.0",
    "fastapi-cache2[memcache] (>=0.2.2,<0.3.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "langchain-ollama (>=0.3.4,<0.4.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-google-genai (>=2.1.7,<3.0.0)",
    "datamodel-code-generator>=0.32.0",
    "matplotlib>=3.10.6",
    "redis>=5.3.1",
    "redis-om>=0.3.5",
    "doppler-env>=0.3.1",
    "alembic>=1.17.2",
    "psycopg2-binary>=2.9.11",
    "asyncpg>=0.31.0",
    "sqlmodel>=0.0.31",
    "beautifulsoup4>=4.13.4",
    "slowapi>=0.1.9",
    "mailchecker>=6.0.1",
]

[project.optional-dependencies]
# uv sync --extra dev
dev = [
    "aiosmtpd (>=1.4.6,<2.0.0)",
    "ruff (>=0.8.1,<1.0.0)"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
podium = "podium.__main__:main"

[tool.ruff]
lint.extend-select= ["FIX002"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.7",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "ty>=0.0.1a25",
    "locust>=2.42.3",
    "textual>=3.0.0",
    "tqdm>=4.67.0",
]

[tool.pytest.ini_options]
addopts = "-p no:warnings"
